backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google AI Integration
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1

# Environment Variables
python-dotenv==1.0.0

# CORS Support
fastapi-cors==0.0.6

# Additional Utilities
pydantic==2.5.0
httpx==0.25.2